+++ libEnigma/src/x86/scoreFunctionAvx2.c
@@ -12,7 +12,7 @@
 #include "computeScoreNoInterleave.h"
 
 // AVX2 scores
-static uint16_t icscoreAvx2( const struct Key* restrict key, scoreLength_t len );
+static int      icscoreAvx2( const struct Key* restrict key, scoreLength_t len );
 static int     uniscoreAvx2( const struct Key* restrict key, scoreLength_t len );
 static int      biscoreAvx2( const struct Key* restrict key, scoreLength_t len );
 static int     triscoreAvx2( const struct Key* restrict key, scoreLength_t len );
@@ -21,7 +21,7 @@ enigma_score_function_t enigmaScoreAvx2 = { triscoreAvx2,  biscoreAvx2 , icscore
 
 union ScoringDecodedMessage decodedMsgPartAvx2;
 
-static uint16_t icscoreAvx2( const struct Key* const restrict key, scoreLength_t len ) {
+static int icscoreAvx2( const struct Key* const restrict key, scoreLength_t len ) {
     DecodeScoredMessagePartAvx2( key, len, &decodedMsgPartAvx2 );
     return ComputeIcscoreFromDecodedMsgAvx2( &decodedMsgPartAvx2, len );
 }
